redoing testing suite setup
