+++ LFPy/test/ball_and_sticks.hoc
