+++ LFPy/test/sinsyn.mod
