+++ LFPy/test/stick.hoc
